fix: fix tests by setting relation tree entities in a new file
