+++ test/github-issues/7155/entity/RemainingTreeEntities.ts
@@ -1,8 +1,6 @@
 import {
     Column,
     Entity,
-    JoinColumn,
-    OneToOne,
     PrimaryColumn,
     PrimaryGeneratedColumn,
     Tree,
@@ -94,79 +92,4 @@ export class MultiIdMaterialized {
 
     @TreeParent({ onDelete: "CASCADE" })
     parent: MultiIdMaterialized | null;
-}
-
-@Entity()
-@Tree("closure-table")
-export class RelationClosure {
-    @PrimaryGeneratedColumn()
-    id: number;
-
-    @TreeChildren()
-    children: RelationClosure[];
-
-    @TreeParent()
-    parent: RelationClosure;
-
-    @OneToOne(() => Relation, { nullable: false })
-    @JoinColumn()
-    relation: Relation;
-
-    @OneToOne(() => OtherRelation, { cascade: true })
-    @JoinColumn()
-    otherRelation: OtherRelation;
-}
-
-@Entity()
-@Tree("nested-set")
-export class RelationNested {
-    @PrimaryGeneratedColumn()
-    id: number;
-
-    @TreeChildren()
-    children: RelationNested[];
-
-    @TreeParent()
-    parent: RelationNested;
-
-    @OneToOne(() => Relation, { nullable: false })
-    @JoinColumn()
-    relation: Relation;
-
-    @OneToOne(() => OtherRelation, { cascade: true })
-    @JoinColumn()
-    otherRelation: OtherRelation;
-}
-
-@Entity()
-@Tree("materialized-path")
-export class RelationMaterialized {
-    @PrimaryGeneratedColumn()
-    id: number;
-
-    @TreeChildren()
-    children: RelationMaterialized[];
-
-    @TreeParent()
-    parent: RelationMaterialized;
-
-    @OneToOne(() => Relation, { nullable: false })
-    @JoinColumn()
-    relation: Relation;
-
-    @OneToOne(() => OtherRelation, { cascade: true })
-    @JoinColumn()
-    otherRelation: OtherRelation;
-}
-
-@Entity()
-export class Relation {
-    @PrimaryGeneratedColumn()
-    id: number;
-}
-
-@Entity()
-export class OtherRelation {
-    @PrimaryGeneratedColumn()
-    id: number;
 }
